Display blockings entry details
Add columns to the blockings table view. There is an important issue with order
of timestamps. Sometimes, events goes backward and produce wrong results. There
is a ugly hack to wait 10sec before launching processing, that highlight the
concurrency problem. If the delay is too low, then the bug occurs.